Move platform-specific pieces of `CertificatePal` into separate classes.
* Pal.OSX/CertificatePal.cs: Move `AppleCertificatePal` into its own file and add
  all static methods from `CertificatePal` to it, turning `CertificatePal` into a
  tiny proxy.

* Pal.Unix/CertificatePal.cs: Rename into `OpenSslCertificatePal` to make it consistent
  with OS X, then add `CertificatePal` proxy.

* Pal.Unix/OpenSslX509Encoder.cs and StorePal.cs: use `OpenSslCertificatePal` instead
  of `CertificatePal`.
