+++ src/System.Security.Cryptography.X509Certificates/src/Internal/Cryptography/Pal.Unix/CertificatePal.cs
@@ -14,190 +14,22 @@ namespace Internal.Cryptography.Pal
     {
         public static ICertificatePal FromHandle(IntPtr handle)
         {
-            if (handle == IntPtr.Zero)
-                throw new ArgumentException(SR.Arg_InvalidHandle, nameof(handle));
-
-            return new OpenSslX509CertificateReader(Interop.Crypto.X509UpRef(handle));
+            return OpenSslX509CertificateReader.FromHandle(handle);
         }
 
         public static ICertificatePal FromOtherCert(X509Certificate cert)
         {
-            Debug.Assert(cert.Pal != null);
-
-            // Ensure private key is copied
-            OpenSslX509CertificateReader certPal = (OpenSslX509CertificateReader)cert.Pal;
-            return certPal.DuplicateHandles();
+            return OpenSslX509CertificateReader.FromOtherCert(cert);
         }
 
         public static ICertificatePal FromBlob(byte[] rawData, SafePasswordHandle password, X509KeyStorageFlags keyStorageFlags)
         {
-            Debug.Assert(password != null);
-
-            ICertificatePal cert;
-            Exception openSslException;
-
-            if (TryReadX509Der(rawData, out cert) ||
-                TryReadX509Pem(rawData, out cert) ||
-                PkcsFormatReader.TryReadPkcs7Der(rawData, out cert) ||
-                PkcsFormatReader.TryReadPkcs7Pem(rawData, out cert) ||
-                PkcsFormatReader.TryReadPkcs12(rawData, password, out cert, out openSslException))
-            {
-                if (cert == null)
-                {
-                    // Empty collection, most likely.
-                    throw new CryptographicException();
-                }
-
-                return cert;
-            }
-
-            // Unsupported
-            Debug.Assert(openSslException != null);
-            throw openSslException;
+            return OpenSslX509CertificateReader.FromBlob(rawData, password, keyStorageFlags);
         }
 
         public static ICertificatePal FromFile(string fileName, SafePasswordHandle password, X509KeyStorageFlags keyStorageFlags)
         {
-            // If we can't open the file, fail right away.
-            using (SafeBioHandle fileBio = Interop.Crypto.BioNewFile(fileName, "rb"))
-            {
-                Interop.Crypto.CheckValidOpenSslHandle(fileBio);
-
-                return FromBio(fileBio, password);
-            }
-        }
-
-        private static ICertificatePal FromBio(SafeBioHandle bio, SafePasswordHandle password)
-        {
-            int bioPosition = Interop.Crypto.BioTell(bio);
-
-            Debug.Assert(bioPosition >= 0);
-
-            ICertificatePal certPal;
-            if (TryReadX509Pem(bio, out certPal))
-            {
-                return certPal;
-            }
-
-            // Rewind, try again.
-            RewindBio(bio, bioPosition);
-
-            if (TryReadX509Der(bio, out certPal))
-            {
-                return certPal;
-            }
-
-            // Rewind, try again.
-            RewindBio(bio, bioPosition);
-
-            if (PkcsFormatReader.TryReadPkcs7Pem(bio, out certPal))
-            {
-                return certPal;
-            }
-
-            // Rewind, try again.
-            RewindBio(bio, bioPosition);
-
-            if (PkcsFormatReader.TryReadPkcs7Der(bio, out certPal))
-            {
-                return certPal;
-            }
-
-            // Rewind, try again.
-            RewindBio(bio, bioPosition);
-
-            // Capture the exception so in case of failure, the call to BioSeek does not override it.
-            Exception openSslException;
-            if (PkcsFormatReader.TryReadPkcs12(bio, password, out certPal, out openSslException))
-            {
-                return certPal;
-            }
-
-            // Since we aren't going to finish reading, leaving the buffer where it was when we got
-            // it seems better than leaving it in some arbitrary other position.
-            // 
-            // Use BioSeek directly for the last seek attempt, because any failure here should instead
-            // report the already created (but not yet thrown) exception.
-            if (Interop.Crypto.BioSeek(bio, bioPosition) < 0)
-            {
-                Interop.Crypto.ErrClearError();
-            }
-
-            Debug.Assert(openSslException != null);
-            throw openSslException;
-        }
-
-        internal static void RewindBio(SafeBioHandle bio, int bioPosition)
-        {
-            int ret = Interop.Crypto.BioSeek(bio, bioPosition);
-
-            if (ret < 0)
-            {
-                throw Interop.Crypto.CreateOpenSslCryptographicException();
-            }
-        }
-
-        internal static bool TryReadX509Der(byte[] rawData, out ICertificatePal certPal)
-        {
-            SafeX509Handle certHandle = Interop.Crypto.DecodeX509(rawData, rawData.Length);
-
-            if (certHandle.IsInvalid)
-            {
-                certHandle.Dispose();
-                certPal = null;
-                Interop.Crypto.ErrClearError();
-                return false;
-            }
-
-            certPal = new OpenSslX509CertificateReader(certHandle);
-            return true;
-        }
-
-        internal static bool TryReadX509Pem(SafeBioHandle bio, out ICertificatePal certPal)
-        {
-            SafeX509Handle cert = Interop.Crypto.PemReadX509FromBio(bio);
-
-            if (cert.IsInvalid)
-            {
-                cert.Dispose();
-                certPal = null;
-                Interop.Crypto.ErrClearError();
-                return false;
-            }
-
-            certPal = new OpenSslX509CertificateReader(cert);
-            return true;
-        }
-
-        internal static bool TryReadX509Pem(byte[] rawData, out ICertificatePal certPal)
-        {
-            using (SafeBioHandle bio = Interop.Crypto.CreateMemoryBio())
-            {
-                Interop.Crypto.CheckValidOpenSslHandle(bio);
-
-                if (Interop.Crypto.BioWrite(bio, rawData, rawData.Length) != rawData.Length)
-                {
-                    Interop.Crypto.ErrClearError();
-                }
-
-                return TryReadX509Pem(bio, out certPal);
-            }
-        }
-
-        internal static bool TryReadX509Der(SafeBioHandle bio, out ICertificatePal fromBio)
-        {
-            SafeX509Handle cert = Interop.Crypto.ReadX509AsDerFromBio(bio);
-
-            if (cert.IsInvalid)
-            {
-                cert.Dispose();
-                fromBio = null;
-                Interop.Crypto.ErrClearError();
-                return false;
-            }
-
-            fromBio = new OpenSslX509CertificateReader(cert);
-            return true;
+            return OpenSslX509CertificateReader.FromFile(fileName, password, keyStorageFlags);
         }
     }
 }
