+++ src/System.Security.Cryptography.X509Certificates/src/Internal/Cryptography/Pal.Unix/StorePal.cs
@@ -31,8 +31,8 @@ namespace Internal.Cryptography.Pal
 
             ICertificatePal singleCert;
 
-            if (CertificatePal.TryReadX509Der(rawData, out singleCert) ||
-                CertificatePal.TryReadX509Pem(rawData, out singleCert))
+            if (OpenSslX509CertificateReader.TryReadX509Der(rawData, out singleCert) ||
+                OpenSslX509CertificateReader.TryReadX509Pem(rawData, out singleCert))
             {
                 // The single X509 structure methods shouldn't return true and out null, only empty
                 // collections have that behavior.
@@ -74,21 +74,21 @@ namespace Internal.Cryptography.Pal
 
             ICertificatePal singleCert;
 
-            if (CertificatePal.TryReadX509Pem(bio, out singleCert))
+            if (OpenSslX509CertificateReader.TryReadX509Pem(bio, out singleCert))
             {
                 return SingleCertToLoaderPal(singleCert);
             }
 
             // Rewind, try again.
-            CertificatePal.RewindBio(bio, bioPosition);
+            OpenSslX509CertificateReader.RewindBio(bio, bioPosition);
 
-            if (CertificatePal.TryReadX509Der(bio, out singleCert))
+            if (OpenSslX509CertificateReader.TryReadX509Der(bio, out singleCert))
             {
                 return SingleCertToLoaderPal(singleCert);
             }
 
             // Rewind, try again.
-            CertificatePal.RewindBio(bio, bioPosition);
+            OpenSslX509CertificateReader.RewindBio(bio, bioPosition);
 
             List<ICertificatePal> certPals;
 
@@ -98,7 +98,7 @@ namespace Internal.Cryptography.Pal
             }
 
             // Rewind, try again.
-            CertificatePal.RewindBio(bio, bioPosition);
+            OpenSslX509CertificateReader.RewindBio(bio, bioPosition);
 
             if (PkcsFormatReader.TryReadPkcs7Der(bio, out certPals))
             {
@@ -106,7 +106,7 @@ namespace Internal.Cryptography.Pal
             }
 
             // Rewind, try again.
-            CertificatePal.RewindBio(bio, bioPosition);
+            OpenSslX509CertificateReader.RewindBio(bio, bioPosition);
 
             // Capture the exception so in case of failure, the call to BioSeek does not override it.
             Exception openSslException;
@@ -263,8 +263,8 @@ namespace Internal.Cryptography.Pal
 
                     ICertificatePal pal;
 
-                    while (CertificatePal.TryReadX509Pem(fileBio, out pal) ||
-                        CertificatePal.TryReadX509Der(fileBio, out pal))
+                    while (OpenSslX509CertificateReader.TryReadX509Pem(fileBio, out pal) ||
+                        OpenSslX509CertificateReader.TryReadX509Der(fileBio, out pal))
                     {
                         X509Certificate2 cert = new X509Certificate2(pal);
